requirements.txt
amqp==5.1.1
anyio==4.0.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
decorator==5.1.1
Django==4.2.2
django-celery-results==2.5.1
django-environ==0.10.0
django-extensions==3.2.3
django-jinja==2.10.2
djangorestframework==3.14.0
easy-thumbnails==2.8.1
executing==1.2.0
generics==6.0.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.3.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-telegram-bot==20.5
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
